Renamed components/common to components/gadgets
