+++ src/main/js/components/gadgets/LoginPopup.js
