+++ src/main/js/components/gadgets/PopupDialog.js
