+++ src/main/js/components/gadgets/SwitchButton.js
